Fix RailsOps.without_authorization
This fixes a bug in without_authorization that, in case of an exception,
prevented setting back the original value that was active before calling
without_authorization. Now it is ensured that authorization is properly
re-activated after calling without_authorization if it was active
before, even if an exception handles in the given block.