Add a `:prepend_action` parameter to the policy function
It is no possible to add a policy in front of the policy chain
instead of the end. It is however not guaranteed to be the first
policy to be run since multiple policies can be prepended to the chain.
A prepended policy can't access the operation model, since it is not set
yet.